test(svelte): extract shared diagnostic mount harness
collect()/settled() move out of wiring-diagnostics.test.ts into
tests/interaction/diagnostic-harness.ts so the upcoming inspect-advisory
mount tests (#1206) reuse the same race-safe primitives instead of
duplicating them. File header documents why these assertions must live
in the browser lane (diagnostic delivery runs in $effect, which
svelte/server never executes).
